+++ sdk/storage/azure_storage_blob/src/lib.rs
@@ -21,4 +21,4 @@ mod logging;
 pub mod models;
 pub use models::error::{Result, StorageError};
 #[cfg(feature = "sas_builder")]
-mod sas_helpers;
+mod sas;