+++ sdk/storage/azure_storage_queue/src/lib.rs
@@ -12,4 +12,7 @@ mod logging;
 pub mod clients;
 pub mod models;
 
+#[cfg(feature = "sas_builder")]
+mod sas;
+
 pub use clients::{QueueClient, QueueClientOptions, QueueServiceClient, QueueServiceClientOptions};